backend/requirements.txt
Authlib==1.2.0
blinker==1.6.2    
certifi==2022.12.7
cffi==1.15.1      
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
distro==1.8.0
Flask==2.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
pycparser==2.21
pypdf==3.8.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
six==1.16.0
tabula-py==2.7.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.16.2
Werkzeug==2.3.2